Conflicts:
	src/action.cc
	src/decaymodes.cc
	src/include/action.h
	src/include/processbranch.h
	src/particletype.cc
	src/scatteraction.cc
	src/tests/processbranch.cc

+++ src/include/resonances.h
@@ -123,7 +123,7 @@ double spectral_function_integrand(double resonance_mass, void * parameters);
  */
 float sample_resonance_mass(const ParticleType &type_resonance,
                             const ParticleType &type_stable,
-                            const float cms_energy);
+                            const double cms_energy);
 
 /**
  * Function for 1-dimensional GSL integration.
